Make random strings more random
`uniqid()` is generated as a function of system time, so it doesn't work
well for generating nonces. The OIDC spec recommends a cryptographically
secure hash of a cryptographically secure string, so I used a sha256
hash of 64 bytes generated by `random_bytes`.

I kept `uniqid()` for LTI_Message_Launch IDs, but because it can lead to
collisions if called twice with the same system clock (possible with
network time resets) I added the `*more_entropy*` parameter to add more
entropy.