Fix invalid request when specifying no custom params
The current default value for $custom_params would be serialized into an array, which would make the request uncompliant to the specification and can lead to issues because the platform expects an object. Casting a null object will instead be serialized into an empty object which it is supposed to be.